# Prealloc in ColStr.DecodeColumn