Try to batch prealloc/avoid append in ColStr.DecodeColumn for better perfomance on many small strings (so memory overhead not too high)
```
$ go test -benchmem -run=^$ -bench ^BenchmarkColStr_DecodeColumn$ github.com/ClickHouse/ch-go/proto

goos: linux
goarch: amd64
pkg: github.com/ClickHouse/ch-go/proto
cpu: Intel(R) Core(TM) i5-9400F CPU @ 2.90GHz
BenchmarkColStr_DecodeColumn-6   	   31533	     38180 ns/op	 864.33 MB/s	       0 B/op	       0 allocs/op
PASS
ok  	github.com/ClickHouse/ch-go/proto	1.596s

$ benchcmp DecodeStr.txt DecodeStr2.txt 
benchmark                          old ns/op     new ns/op     delta
BenchmarkColStr_DecodeColumn-6     41783         38180         -8.62%

benchmark                          old MB/s     new MB/s     speedup
BenchmarkColStr_DecodeColumn-6     789.80       864.33       1.09x

benchmark                          old allocs     new allocs     delta
BenchmarkColStr_DecodeColumn-6     0              0              +0.00%

benchmark                          old bytes     new bytes     delta
BenchmarkColStr_DecodeColumn-6     0             0             +0.00%

```