@msaf1980 2022-12-08
It's efficient for fetch many small strings. This is a fairly frequent case. Not edge.