@msaf1980 2022-12-08
If needed, i can write more benchmark, for see detailed effect on many cases.